Explosion terminada

+++ skeleton/GravityGenerator.cpp
@@ -1,6 +1,6 @@
 #include "GravityGenerator.h"
 
-GravityGenerator::GravityGenerator(ParticleSystem* pool, Vector3 direction, double gravity)
+GravityGenerator::GravityGenerator(ParticleSystem* pool, Vector3 direction, float gravity)
 	: ForceGenerator(pool) {
 	this->direction = direction;
 	this->gravity = gravity;
