+++ skeleton/WindGenerator.h
@@ -4,11 +4,11 @@
 class WindGenerator : public ForceGenerator
 {
 public:
-	WindGenerator(ParticleSystem* sysRef, Vector3 direction, double radius, double strength);
+	WindGenerator(ParticleSystem* sysRef, Vector3 direction, float radius, float strength);
 
 	virtual void applyForce(Particle* p) override;
 private:
 	Vector3 direction;
-	double radius;
-	double strength;
+	float radius;
+	float strength;
 };
